requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
pypdf>=3.14.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
numpy>=1.26.0
requests>=2.31.0
charset-normalizer>=3.3.2
beautifulsoup4